TC 3561: Fix an infinite loop TC 6304: Prevent MoveFollow and MoveChase while UNIT_FLAG_DISABLE_MOVE is present TC 2439: Set default movement as follow movement for charmed/pet when enterevademode

+++ src/game/ArenaTeamHandler.cpp
@@ -28,7 +28,6 @@
 #include "Log.h"
 #include "ObjectMgr.h"
 #include "SocialMgr.h"
-#include "Language.h"
 
 void WorldSession::HandleInspectArenaStatsOpcode(WorldPacket & recv_data)
 {
@@ -301,7 +300,7 @@ void WorldSession::HandleArenaTeamRemoveFromTeamOpcode(WorldPacket & recv_data)
     MapEntry const* mapEntry = sMapStore.LookupEntry(player->GetMapId());
 
     if (mapEntry && mapEntry->IsBattleArena())
-           return;
+        return;
 
     at->DelMember(member->guid);
 
@@ -355,10 +354,10 @@ void WorldSession::HandleArenaTeamPromoteToCaptainOpcode(WorldPacket & recv_data
 void WorldSession::SendArenaTeamCommandResult(uint32 team_action, const std::string& team, const std::string& player, uint32 error_id)
 {
     WorldPacket data(SMSG_ARENA_TEAM_COMMAND_RESULT, 4+team.length()+1+player.length()+1+4);
-    data << team_action;
+    data << uint32(team_action);
     data << team;
     data << player;
-    data << error_id;
+    data << uint32(error_id);
     SendPacket(&data);
 }
